feat(react): bump wayfinder-react exports, update docs
